Introduce the junction point concept to fix #5
Since Vista the IPersistFolder2::GetCurFolder should return the actual folder. GetCurFolder would however always return the path the junction point. This prevents that a user could go into a sub-directory.
The introduced junction point member maintains the ITEMIDLIST to the original location in the namespace. The m_pidlFolder member if the absolute path to the current folder.